explicitly handle imagemagick exceptions by not attempting to save image
